pyproject.toml
[project]
name = "chodu"
version = "0.1.0"
description = "A project using FastAPI, LangChain, and more"
authors = [
    { name="Kartik"}
]
requires-python = ">=3.8"

dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "faiss-cpu",
    "pymupdf",
    "python-docx",
    "python-dotenv",
    "requests",
    "rank_bm25"
]

[tool.uv]
# Optional: uv-specific settings can go here

[build-system]
requires = ["uv"]
build-backend = "uv.build"